# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RecoveryOS API",
    "description": "Auditable subscription-recovery orchestration.",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/demo/fixtures/{fixture_name}": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "Get Demo Fixture",
        "description": "Return a versioned, synthetic screen fixture with no provider side effects.",
        "operationId": "getDemoFixture",
        "parameters": [
          {
            "name": "fixture_name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "dashboard",
                "case-detail",
                "customer-agent",
                "customer-voice",
                "ml-lab"
              ],
              "type": "string",
              "title": "Fixture Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getdemofixture"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}